Reorganize database commands in separate files.
